// PM_Case_Management_2/client/src/app/pages/PM/progress-report/progress-report/Iprogress-report.ts
export interface FilterationCriteria {

   budgetYear: number
   empId?: string
   planId?: string
   taskId?: string
   actParentId?: string
   actId?: string
   reporttype: string


   structureId?: string
   filterbyId?: number

   Quarter?: string
   Month?: string
   FromDate?: string
   ToDate?: string

}